Blacklists IO::E* classes
I'm removing these for now. Looks like they are dynamically generated and inherit from `Errno::E*` classes that, for whatever reason, are not listed as Ruby core. (It looks like they are generated straight from the C source, but I'll have to research more.) Regenerated all data files in light of this.